add annealer.cpp/.h
