+++ vpr/src/base/read_options.cpp
@@ -3141,13 +3141,13 @@ void set_conditional_defaults(t_options& args) {
         || args.PlaceAlphaDecay.provenance() == Provenance::SPECIFIED
         || args.PlaceSuccessMin.provenance() == Provenance::SPECIFIED
         || args.PlaceSuccessTarget.provenance() == Provenance::SPECIFIED) {
-        args.anneal_sched_type.set(DUSTY_SCHED, Provenance::INFERRED);
+        args.anneal_sched_type.set(e_sched_type::DUSTY_SCHED, Provenance::INFERRED);
     } else if (args.PlaceInitT.provenance() == Provenance::SPECIFIED // Any of these flags select a manual schedule
                || args.PlaceExitT.provenance() == Provenance::SPECIFIED
                || args.PlaceAlphaT.provenance() == Provenance::SPECIFIED) {
-        args.anneal_sched_type.set(USER_SCHED, Provenance::INFERRED);
+        args.anneal_sched_type.set(e_sched_type::USER_SCHED, Provenance::INFERRED);
     } else {
-        args.anneal_sched_type.set(AUTO_SCHED, Provenance::INFERRED); // Otherwise use the automatic schedule
+        args.anneal_sched_type.set(e_sched_type::AUTO_SCHED, Provenance::INFERRED); // Otherwise use the automatic schedule
     }
 
     /*
